29 Nov 2022

+++ Bab 1/Tugas/2/persegi-panjang.cpp
@@ -38,7 +38,7 @@ main()
     }
     else if (jenis == 'd')
     {
-        int diagonal = sqrt((panjang * panjang) + (lebar * lebar));
+        float diagonal = sqrt((panjang * panjang) + (lebar * lebar));
         cout << "Diagonal = " << diagonal;
     }
 }
